requirements.txt
aiodns @ file:///opt/miniconda3/conda-bld/aiodns_1762184800293/work
aiohappyeyeballs @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_86beu757ao/croot/aiohappyeyeballs_1755769847762/work
aiohttp @ file:///Users/ec2-user/croot/aiohttp_1767813704691/work
aiosignal @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_e3cjtbh902/croot/aiosignal_1755870814211/work
anyio @ file:///opt/miniconda3/conda-bld/anyio_1758623422202/work
appnope @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_46iquy4y0r/croot/appnope_1750774774351/work
asgiref @ file:///opt/miniconda3/conda-bld/asgiref_1760342044893/work
asttokens @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_11i8cnwuxs/croot/asttokens_1743630449718/work
attrs @ file:///opt/miniconda3/conda-bld/attrs_1762356931094/work
Bottleneck @ file:///opt/miniconda3/conda-bld/bottleneck_1761937852211/work
brotlicffi @ file:///opt/miniconda3/conda-bld/brotlicffi_1764961161081/work
certifi @ file:///Users/ec2-user/croot/certifi_1767659143057/work/certifi
cffi @ file:///opt/miniconda3/conda-bld/cffi_1761832773185/work
charset-normalizer @ file:///opt/miniconda3/conda-bld/charset-normalizer_1761744840553/work
comm @ file:///opt/miniconda3/conda-bld/comm_1763119153598/work
contourpy==1.3.3
cycler==0.12.1
datasets @ file:///Users/ec2-user/croot/datasets_1766497781450/work
debugpy @ file:///opt/miniconda3/conda-bld/debugpy_1762421637290/work
decorator @ file:///opt/miniconda3/conda-bld/decorator_1757341238163/work
dill @ file:///opt/miniconda3/conda-bld/dill_1760096784777/work
Django @ file:///opt/miniconda3/conda-bld/django_1764329190926/work
executing @ file:///opt/miniconda3/conda-bld/executing_1757061235986/work
filelock @ file:///Users/ec2-user/croot/filelock_1769440497771/work
fonttools==4.61.1
frozenlist @ file:///opt/miniconda3/conda-bld/frozenlist_1761062014412/work
fsspec @ file:///opt/miniconda3/conda-bld/fsspec_1762432466128/work
h11 @ file:///opt/miniconda3/conda-bld/h11_1761931258823/work
hf-xet @ file:///Users/ec2-user/croot/hf-xet_1769676990399/work
httpcore @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_941yf7p5zt/croot/httpcore_1748526061031/work
httpx @ file:///opt/miniconda3/conda-bld/httpx_1760447187686/work
huggingface_hub @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_29dc5yixdo/croot/huggingface_hub_1755770856782/work
idna @ file:///opt/miniconda3/conda-bld/idna_1761912005426/work
ipykernel @ file:///opt/miniconda3/conda-bld/ipykernel_1762946073576/work
ipython @ file:///opt/miniconda3/conda-bld/ipython_1762789298907/work
ipython_pygments_lexers @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_25fteymumy/croot/ipython_pygments_lexers_1744753256449/work
jedi @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_38ctoinnl0/croot/jedi_1733987402850/work
jupyter_client @ file:///Users/ec2-user/croot/jupyter_client_1768399009041/work
jupyter_core @ file:///opt/miniconda3/conda-bld/jupyter_core_1764588283979/work
kiwisolver==1.4.9
matplotlib==3.10.8
matplotlib-inline @ file:///opt/miniconda3/conda-bld/matplotlib-inline_1762779171616/work
multidict @ file:///Users/ec2-user/croot/multidict_1770629142335/work
multiprocess @ file:///Users/ec2-user/croot/multiprocess_1766144193974/work
nest_asyncio @ file:///Users/builder/cbouss/perseverance-python-buildout/croot/nest-asyncio_1728588080290/work
numexpr @ file:///opt/miniconda3/conda-bld/numexpr_1762165518100/work
numpy @ file:///Users/ec2-user/croot/numpy_and_numpy_base_1770732574516/work/dist/numpy-2.4.2-cp313-cp313-macosx_12_0_arm64.whl#sha256=d3095bea5e8f18db16f93173963cea2f153b6589580537735755c455bb529958
packaging @ file:///opt/miniconda3/conda-bld/packaging_1761049079023/work
pandas @ file:///opt/miniconda3/conda-bld/pandas_1762332179809/work/dist/pandas-2.3.3-cp313-cp313-macosx_12_0_arm64.whl#sha256=e15ab5e76dfcac9817c2633549d6e4cb87e6a735ade38ff7afbfa488f280b9ba
parso @ file:///opt/miniconda3/conda-bld/parso_1762781703500/work
pexpect @ file:///opt/miniconda3/conda-bld/pexpect_1762535936187/work
pillow==12.1.1
platformdirs @ file:///opt/miniconda3/conda-bld/platformdirs_1762356666077/work
prompt_toolkit @ file:///opt/miniconda3/conda-bld/prompt-toolkit_1761744878191/work
propcache @ file:///opt/miniconda3/conda-bld/propcache_1765457264444/work
psutil @ file:///opt/miniconda3/conda-bld/psutil_1761896323233/work
ptyprocess @ file:///opt/miniconda3/conda-bld/ptyprocess_1762424170819/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl#sha256=3470be7f810474c8a2ecfcd6e02acc6aea8483ab595417fa4e336362a349933e
pure_eval @ file:///opt/miniconda3/conda-bld/pure_eval_1757067068750/work
pyarrow @ file:///Users/ec2-user/croot/pyarrow_1771937513883/work/python
pycares @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_b01dqj7707/croot/pycares_1756217969873/work
pycparser @ file:///opt/miniconda3/conda-bld/pycparser_1757496048647/work
Pygments @ file:///opt/miniconda3/conda-bld/pygments_1762431584667/work
pyparsing==3.3.2
PySocks @ file:///opt/miniconda3/conda-bld/pysocks_1761753010927/work
python-dateutil @ file:///Users/builder/cbouss/perseverance-python-buildout/croot/python-dateutil_1728585579455/work
python-decouple==3.8
pytz @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_3bjczh6it0/croot/pytz_1752135865541/work
PyYAML @ file:///opt/miniconda3/conda-bld/pyyaml_1763465492884/work
pyzmq @ file:///opt/miniconda3/conda-bld/pyzmq_1762375585960/work
regex @ file:///opt/miniconda3/conda-bld/regex_1765378749968/work
requests @ file:///opt/miniconda3/conda-bld/requests_1762359592010/work
safetensors @ file:///opt/miniconda3/conda-bld/safetensors_1762515995319/work
setuptools==80.9.0
six @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_03myqm7p6o/croot/six_1744271511946/work
sniffio @ file:///opt/miniconda3/conda-bld/sniffio_1764329150559/work
sqlparse @ file:///opt/miniconda3/conda-bld/sqlparse_1764240029253/work
stack_data @ file:///opt/miniconda3/conda-bld/stack_data_1757067040463/work
tokenizers @ file:///Users/ec2-user/croot/tokenizers_1767006646821/work/bindings/python
tornado @ file:///Users/ec2-user/croot/tornado_1765992646537/work
tqdm @ file:///Users/ec2-user/croot/tqdm_1771398699921/work
traitlets @ file:///Users/builder/cbouss/perseverance-python-buildout/croot/traitlets_1728585482264/work
transformers @ file:///opt/miniconda3/conda-bld/transformers_1763719387611/work
typing_extensions @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_dffmcwqo6o/croot/typing_extensions_1756281471323/work
tzdata @ file:///home/task_176829656558347/croot/python-tzdata_1768296637279/work
urllib3 @ file:///Users/ec2-user/croot/urllib3_1767810404629/work
uvloop @ file:///opt/miniconda3/conda-bld/uvloop_1762175502840/work
wcwidth @ file:///Users/ec2-user/croot/wcwidth_1767780816236/work
wheel==0.45.1
whitenoise==6.11.0
xxhash @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_4dqzb3ngyc/croot/python-xxhash_1737039922516/work
yarl @ file:///opt/miniconda3/conda-bld/yarl_1762250141518/work
